Update the codes to make more modular
