+++ src/components/crud.js
@@ -1,6 +1,7 @@
 import { inputTaker, lists, warning } from "./variables";
 import { renderAll } from "../index";
 import { storeToLocal } from "./storage";
+
 import { tasks } from "../index";
 const saveToData = (event) => {
   event.preventDefault();
@@ -20,9 +21,9 @@ const saveToData = (event) => {
   }
 };
 const clearSpace = () => {
-    inputTaker.value = "";
+  inputTaker.value = "";
 };
-const deleteTask=(event) => {
+const deleteTask = (event) => {
   if (event.target.classList.contains("delete-btn")) {
     const index = event.target.dataset.index;
     tasks.splice(index, 1);
@@ -30,11 +31,33 @@ const deleteTask=(event) => {
     tasks.forEach((todo, index) => {
       todo.index = index + 1;
     });
-    storeToLocal(tasks)
+    storeToLocal(tasks);
     renderAll();
-
-    
   }
+};const EditTask = (index) => {
+  const task = tasks[index];
+  const descriptionElem = document.querySelector(
+    `.task:nth-of-type(${index + 1}) .input-board`
+  );
+  console.log(descriptionElem);
+
+  const inputElem = document.createElement("input");
+  inputElem.type = "text";
+  inputElem.value = task.description;
+
+  const saveBtn = document.createElement("button");
+  saveBtn.textContent = "Save";
+  saveBtn.classList.add('save');
+  saveBtn.addEventListener("click", () => {
+    tasks[index].description = inputElem.value;
+    descriptionElem.textContent = inputElem.value;
+    renderAll();
+    storeToLocal(tasks);
+  });
+
+  const divElem = document.createElement("div");
+  divElem.append(inputElem, saveBtn);
+  descriptionElem.replaceWith(divElem);
 };
 
-export {deleteTask ,saveToData};
+export { deleteTask, saveToData,EditTask };
