+++ src/index.js
@@ -3,6 +3,7 @@ import { saveToData, deleteTask } from "./components/crud";
 import { form, inputTaker, deleteBtn } from "./components/variables";
 import { lists } from "./components/variables";
 import { retrieveFromLocal, storeToLocal } from "./components/storage";
+import { EditTask } from "./components/crud";
 let tasks;
 if (retrieveFromLocal()) {
   tasks = retrieveFromLocal();
@@ -46,31 +47,10 @@ const renderAll = () => {
     lists.append(hr);
   });
 };
+const handleEdit=(index)=>{
+  EditTask(index);
+}
 
-const handleEdit = (index) => {
-  const task = tasks[index];
-  const descriptionElem = document.querySelector(
-    `.task:nth-of-type(${index + 1}) .input-board`
-  );
-  console.log(descriptionElem);
 
-  const inputElem = document.createElement("input");
-  inputElem.type = "text";
-  inputElem.value = task.description;
-
-  const saveBtn = document.createElement("button");
-  saveBtn.textContent = "Save";
-  saveBtn.classList.add('save');
-  saveBtn.addEventListener("click", () => {
-    tasks[index].description = inputElem.value;
-    descriptionElem.textContent = inputElem.value;
-    renderAll();
-    storeToLocal(tasks);
-  });
-
-  const divElem = document.createElement("div");
-  divElem.append(inputElem, saveBtn);
-  descriptionElem.replaceWith(divElem);
-};
 
 export { tasks, renderAll };
